pyproject.toml
[project]
name = "magic-interview-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "python-dotenv>=1.1.1",
    "supabase>=2.16.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]
